Initial Commit
`psy-x` is a library for reading and writing Psy-Q LIB and OBJ files.
It's goal is to provide an easy way to read and process LIB and OBJ
files, and maybe one day it will do many of the things that DUMPOBJ.EXE
and PSYLIB.EXE did.

It can currently roundtrip nearly all of the OBJ and LIB files from
various versions of the PS1 & Saturn SDKs. Of the files that cannot be
read, several are actually COFF files and one of the sample OBJs has an
additional record that DUMPOBJ.EXE cannot read, either.

When dumping LIB or OBJ files, the output is roughly equivalent to the
output of PSYLIB.EXE and DUMPOBJ.EXE respectively.

`psyx` supports the `split` command which will extract OBJs from their
containing `LIB`.